fix(compiler): compiler panic when calling namespaced function within index expression (#6882)
fixes #6661 
The `TypeReferenceTransformer` uses a fold visitor pattern to visit certain `InstanceMember` references and convert them into `TypeReference` if needed.
The other arm stopped the visiting recursion by just returning the node and not a call to the folder- therefore inner references were never visited. It was fixed by calling the fold visitor at all arms of the reference enum.

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.