fix(simulator): lockfile race condition
In some cases, the lockfile would try to update the lockfile utimes after it was released due to the async nature of the implementation. This changeset refactors the implementation so it's purely sync.

In the future, the implementation could be refactored into async again, but the release method should wait for the lockfile update before actually releasing the lockfile (if it's actually happening during the release call).