fix(compiler): cannot lift a class that extends a cloud resource (#6490)
This PR aims to fix an issue where unexpected bugs happen when you try extending a class from the Wing SDK, like `cloud.Bucket`. The root cause is related to how Wings' inheritance is implemented in JavaScript.

Each cloud resource has a JavaScript class modeling its inflight API. The class needs to be constructed with target-specific arguments. For example, for the implementation of `cloud.Bucket`, there's a class definition `BucketClient` for each of the supported cloud targets (`tf-aws`, `tf-gcp`, and `tf-azure`). What's important to note is that each class needs to be instantiated at runtime with value that are specific to that cloud target. For example, `new aws.BucketClient(...)` expects an S3 Bucket name, while `new azure.BucketClient(...)` expects an Azure container name _and_ an Azure storage account name.

How does Wing ensure the right values are passed when the class is constructed? Currently, each preflight class has an `_toInflight()` method which returns a string of JavaScript code. That code instantiates the class and hardcodes in the necessary arguments. This method is code-generated by the compiler for most Wing classes.

But suppose we have a class `MyBucket` that extends `Bucket`. To instantiate such a class, we'd need to instantiate `MyBucket` with both the arguments needed by `MyBucket` AND the arguments needed by `Bucket`, and ensure that the `super()` call in `MyBucket`'s constructor passes the appropriate arguments to `Bucket`'s constructor. 

To support this, we're now exposing a new method in the TypeScript source code named `_liftedState()` which can be overridden by classes inside our TypeScript SDK implementation to specify what arguments they need when they're instantiated. When `_toInflight()` is called, it will automatically call `_liftedState()` to obtain the constructor arguments _and_ instantiate the class. But if you just call `_toInflightType()`, the class will be returned without instantiating an instance of it. These internal changes should not have any user-facing effects.

Fixes #4203
Fixes #6093
Fixes #6851

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.