fix(sdk): more resources are created than needed for tests (#6948)
Fixes https://github.com/winglang/wing/issues/6316

This PR addresses an issue where if you just one to run an individual test (in the CLI or Wing Console), it simulates more resources than are needed, which can have side effects like excess Docker containers being spun up on the developer's machine.

The root cause of the issue is that to ensure test isolation, the Wing source file(s) are compiled into a Wing simulator file with multiple copies of the users app (one for each test defined in the entire app). But this assumes all of the tests are going to be executed. If only one or a handful of tests are to be executed, we don't need all of those environments.

As a solution, this PR changes the testing mechanism so a compilation for the simulator always produces a single copy of the app, and when tests are executed, each test is run sequentially, stopping and restarting the simulator as needed.

## Checklist

- [ ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [ ] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.