docs(docs): added examples into docs folder (#6956)
Working on a new `/examples page` which you can see here (https://github.com/winglang/docsite/pull/957) and a [live preview here](https://wing-docs-git-new-examples-page-monada.vercel.app/docs/examples).

The examples that power this page will live in this project (we can always move in the future if it becomes too much!).
Any code snippet marked with `example` will be compiled and checked using Wing to verify the code works (this is already in place for this repo).

The docsite pulls in these examples and renders them on the website.

## Checklist

- [x ] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.