fix(console)!: not all diagnostics are errors (#6975)
The compile function from the @winglang/compiler package throws even if the compilation was successful (e.g. the compiler emitted a warning). As result of that behavior, the console shows a blue screen of death when the compiler emits a warning.

This changeset refactors compile so it doesn't throw anymore, but returns the output directory (if possible) along with the diagnostics and any preflight errors.

Co-authored-by: Cristian Pallarés <cristianp@wing.cloud>