feat: `@type` intrinsic (#7151)
Closes #7150

Adds for the new `@type` intrinsic function. This function can be passed a type, and it will give you back an object (`std.reflect.Type`) with information about it -- for example, whether it's a struct or class, what properties or fields it has, etc. You can use this to validate request data structures at runtime, generate schemas for APIs or databases, or for simply asserting on the structure of your code.

Future work:
- Support doc reflection
- Fix support for variadic functions
- Open discussion about accessing private fields or methods via reflection
- Open issue about explicit "preflight" annotation
- Open issue about explicit `TypeId` enum

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.