chore(docs): moving typescript docs into own section (#7222)
The docs are getting re-looked and updated slightly to help with the flow. Part of this is moving TypeScript docs out of the main project flow (for now), this will create a new section `/docs/typescript` and things will continue to work, but the main project will no longer point to these files, but will still be indexed.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.